Credit fool

See merge request enigma-camp/malang-branch/torvalds-batch-4/team/final-projects/team-4/giri-guide-backend!15

+++ src/main/java/com/abdav/giri_guide/model/response/TourGuideDetailResponse.java
@@ -13,10 +13,10 @@ public record TourGuideDetailResponse(
         Integer totalReview,
         Integer totalCustomer,
 
-        Double price,
-        Double additionalPrice,
+        Long price,
+        Long additionalPrice,
         Integer totalPorter,
-        Double pricePorter,
+        Long pricePorter,
 
         List<MountainListHikingPointResponse> mountains
